fix category product
